Fix: Sonatype Nexus publishing
There seems to be a problem with `group` and `version` set on the root project conflicting with the same set on the subprojects when trying to use the Nexus publishing plugin: https://github.com/gradle-nexus/publish-plugin/issues/150

These changes resolve this by aligning all of them. Once this needs to deviate we may have to revisit.