Prepare `IntervalUnion` for storing interval trees
The concrete `MutableIntervalUnion` implementation is replaced with a `EmptyIntervalUnion` (closes #28) and `IntervalUnionPair`, the latter of which always holds exactly two other non-empty intervals. Combined with `Interval` (which is also a union), these can serve as nodes in an interval tree.

As part of this, `getBounds()` is redefined to always return intervals in canonical form. Not only is this a more well-defined contract, it also prevents having to canonicalize or call `lowerBound/upperBound` within `IntervalUnion`; `start` and `end` can safely be used.